move useDisconnectButton hook
