+++ packages/react/src/components/controls/DisconnectButton.tsx
@@ -1,31 +1,9 @@
-import { ConnectionState } from 'livekit-client';
-import { setupDisconnectButton } from '@livekit/components-core';
 import * as React from 'react';
-import { useRoomContext } from '../../context';
-import { mergeProps } from '../../utils';
-import { useConnectionState } from '../../hooks';
+import { useDisconnectButton } from '../../hooks';
 
 /** @public */
 export interface DisconnectButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
   stopTracks?: boolean;
-}
-
-/** @public */
-export function useDisconnectButton(props: DisconnectButtonProps) {
-  const room = useRoomContext();
-  const connectionState = useConnectionState(room);
-
-  const buttonProps = React.useMemo(() => {
-    const { className, disconnect } = setupDisconnectButton(room);
-    const mergedProps = mergeProps(props, {
-      className,
-      onClick: () => disconnect(props.stopTracks ?? true),
-      disabled: connectionState === ConnectionState.Disconnected,
-    });
-    return mergedProps;
-  }, [room, props, connectionState]);
-
-  return { buttonProps };
 }
 
 /**
